requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
backoff==2.2.1
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
chardet==5.2.0
charset-normalizer==3.4.2
click==8.0.3
cloud-init==24.4
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
dotenv==0.9.9
exceptiongroup==1.3.0
greenlet==3.2.3
grpcio==1.73.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.0.3
jiter==0.10.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.0.1
mem0ai==0.1.114
more-itertools==8.10.0
netifaces==0.11.0
numpy==2.2.6
oauthlib==3.2.0
openai==1.93.2
portalocker==2.10.1
posthog==6.0.3
protobuf==6.31.1
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycurl==7.44.1
pydantic==2.11.7
pydantic_core==2.33.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
PyMySQL==1.1.1
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==5.4.1
qdrant-client==1.14.3
requests==2.32.4
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.41
systemd-python==234
tqdm==4.67.1
Twisted==22.1.0
typing-inspection==0.4.1
typing_extensions==4.14.1
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.5.0
wadllib==1.3.6
zipp==1.0.0
zope.interface==5.4.0
